reorganize directory so that Routes are separate from App
