+++ src/components/index.js
@@ -1,201 +1,5 @@
-import React, { Component } from "react";
-import { 
-  Route, Router, Redirect, Switch 
-} from "react-router-dom";
-import {
-  CheckIn, OpenTabs, BottomNavigationBar,
-  SplashScreen, CustomerInfo, Menu, Logo,
-} from "./customer";
-import { Login, Register } from "./auth";
-import { logout } from "../helpers";
-import { firebaseAuth } from "../config";
-import { AppBar, FlatButton } from "material-ui";
-import { fetchAllMerchants } from "../helpers";
-import { AdminHome, AdminSingleTab } from "./admin";
-import history from "./history";
-
-function PrivateRoute({
-  component: Component,
-  authed,
-  allOpenMerchants,
-  ...rest
-}) {
-  return (
-    <Route
-      {...rest}
-      render={props =>
-        authed === true ? (
-          <Component allOpenMerchants={allOpenMerchants} {...props} />
-        ) : (
-          <Redirect
-            to={{ pathname: "/login", state: { from: props.location } }}
-          />
-        )
-      }
-    />
-  );
-}
-
-function PublicRoute({ component: Component, authed, ...rest }) {
-  return (
-    <Route
-      {...rest}
-      render={props =>
-        authed === false ? <Component {...props} /> : <Redirect to="/" />
-      }
-    />
-  );
-}
-
-export default class App extends Component {
-  state = {
-    authed: false,
-    loading: true,
-    allOpenMerchants: [],
-    openMenu: false
-  };
-  async componentDidMount() {
-    this.removeListener = firebaseAuth().onAuthStateChanged(user => {
-      if (user) {
-        this.setState(_ => ({
-          authed: true,
-          loading: false
-        }));
-      } else {
-        this.setState(_ => ({
-          authed: false,
-          loading: false
-        }));
-      }
-    });
-    const allOpenMerchants = await fetchAllMerchants();
-    this.setState(_ => ({
-      allOpenMerchants
-    }));
-  }
-  componentWillUnmount() {
-    this.removeListener();
-  }
-
-  toggleMenu = _ => this.setState(_ => ({ openMenu: !this.state.openMenu }));
-
-  handleClose = _ => this.setState(_ => ({ openMenu: false }));
-
-  render() {
-    const authButtons = this.state.authed ? (
-      <FlatButton
-        label="Logout"
-        onClick={() => {
-          logout();
-        }}
-        style={{ color: "#fff" }}
-      />
-    ) : null;
-
-    const topbarButtons = <div>{authButtons}</div>;
-
-    const tabData = {
-      0: {
-        path: this.state.authed ? "/" : "/login",
-        icon: this.state.authed ? "plus-circle" : "sign-in",
-        label: this.state.authed ? "New Tab" : "Login"
-      },
-      1: {
-        path: this.state.authed ? "/open-tabs" : "/register",
-        icon: this.state.authed ? "sticky-note" : "user-plus",
-        label: this.state.authed ? "Open Tabs" : "Register"
-      }
-    };
-
-    const { allOpenMerchants, authed } = this.state;
-
-    return this.state.loading === true && allOpenMerchants.length === 0 ? (
-      <SplashScreen />
-    ) : (
-      <Router history={history}>
-        <div className="fullheight">
-          <div>
-            <AppBar
-              title={<Logo />}
-              iconElementRight={topbarButtons}
-              iconStyleRight={{
-                display: "flex",
-                alignItems: "center",
-                marginTop: "0",
-              }}
-              style={{
-                height: '64px',
-                flexShrink: 0,
-                background: 'linear-gradient(to bottom right, #0a2009, #0d2d0b)'
-              }}
-              onLeftIconButtonTouchTap={this.toggleMenu}
-            />
-            <Menu 
-              user={authed}
-              toggleMenu={this.toggleMenu} 
-              openMenu={this.state.openMenu}
-              handleClose={this.handleClose}
-            />
-            <div className="splashbox"></div>
-          </div>
-          <div className="container d-flex justify-content-center mt-3 scrollable">
-            <div className="row">
-              <Switch>
-                <PrivateRoute
-                  path="/"
-                  exact
-                  authed={this.state.authed}
-                  allOpenMerchants={allOpenMerchants}
-                  component={CheckIn}
-                />
-                <PrivateRoute
-                  path="/payment-details"
-                  exact
-                  authed={this.state.authed}
-                  component={CustomerInfo}
-                />
-                <PublicRoute
-                  authed={this.state.authed}
-                  path="/login"
-                  component={Login}
-                />
-                <PublicRoute
-                  authed={this.state.authed}
-                  path="/register"
-                  component={Register}
-                />
-                <PrivateRoute
-                  authed={this.state.authed}
-                  exact
-                  path="/open-tabs"
-                  component={OpenTabs}
-                />
-                <PrivateRoute
-                  authed={this.state.authed}
-                  exact
-                  path="/open-tabs/:id"
-                  component={OpenTabs}
-                />
-                <PrivateRoute
-                  authed={this.state.authed}
-                  exact
-                  path="/admin"
-                  component={AdminHome}
-                />
-                <PrivateRoute
-                  authed={this.state.authed}
-                  exact
-                  path="/admin/tabs/:tabId"
-                  component={AdminSingleTab}
-                />
-                <Route render={_ => <h3>No Match</h3>} />
-              </Switch>
-            </div>
-          </div>
-          <BottomNavigationBar data={tabData} />
-        </div>
-      </Router>
-    );
-  }
-}
-
+export * from './admin';
+export * from './auth';
+export * from './customer';
+export {default as history} from './history';
+export * from './main';
