+++ src/components/main/index.js
@@ -1,0 +1,2 @@
+export { default as App } from "./App";
+export { default as Routes } from "./Routes";
